systemd/adapt: add explicit assertion that we called sd_event_add_time() with a @source output argument
Systemd supports omiting the output source argument. In this case,
the created event source is floating and the reference count
is handled properly.

There are however now callers that make use of that, so instead
of implementing it, must assert against it.

This isn't a change in behavior, because previously we would SEGFAULT.