Modify NMDeviceModem's available logic
This patch modifies NMDeviceModem's available logic such that
the device is only considered available if the modem_state is
>= NM_MODEM_STATE_REGISTERED.  NMDevice defines 'available' as
meaning the device is in such a state that it can be activated.
This change prevents NM from trying to activate a modem which
is not yet ready to be activated.

Bug-Ubuntu: https://bugs.launchpad.net/bugs/1445080

====

LR FIXED!!!